Add libazdice.
Add a parser (and roller) for various dice strings, which supports addition and subtraction of multiple dice groups with modifiers such as minimum value per dice, maximum value per dice, re-rolls for groups and explosive dice. It also supports generation of probability distributions and has very simple API (incomplete) for use in C/C++. 
NB- I don't know whether this is a little bit trivial for `rust-awesome` but the dice parser on roll20 fails to impress.